secure/osin/ingestion/requirements.txt
kafka-python
snscrape
pyyaml
asyncio
praw
flask
google-api-python-client
feedparser
requests
prometheus_client
